IMP-1539 - New label set up

+++ dtos/src/main/java/mobi/nowtechnologies/server/shared/dto/PaymentPolicyDto.java
@@ -17,7 +17,7 @@ public class PaymentPolicyDto {
 
     public static final String PAYMENT_POLICY_DTO = "paymentPolicy";
 
-    private short id;
+    private Integer id;
     private BigDecimal subcost;
     private Integer subweeks;
     private Integer operator;
@@ -66,10 +66,10 @@ public class PaymentPolicyDto {
 		return subweeks == null || subweeks.intValue() == 0;
 	}
     
-    public short getId() {
+    public Integer getId() {
         return id;
     }
-    public void setId(short id) {
+    public void setId(Integer id) {
         this.id = id;
     }
     public BigDecimal getSubcost() {
